# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for transport booking",
        "title": "SimbirGO REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alexander Soldatov",
            "email": "soldatovalex207z@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
...
        "/api/Payment/Hesoyam/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Пополняет баланс пользователся с id = {id} на 250 000.\nАдминистраторы могут пополнять баланс любому пользователю, обычные пользователи только себе.",
                "tags": [
                    "PaymentController"
                ],
                "summary": "Пополнение баланса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpUtil.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpUtil.ResponseError": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string",
                    "example": "error occures"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}